requirements.txt
# ── Backend ────────────────────────────────────────────────────────────────
tensorflow>=2.13.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.9.0

# ── Frontend ───────────────────────────────────────────────────────────────
gradio>=4.28.0
requests>=2.31.0
